monticore-grammar/src/test/java/de/monticore/prettyprint/CommonExpressionsPrettyPrinterTest.java
/* (c) https://github.com/MontiCore/monticore */

package de.monticore.prettyprint;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import java.io.IOException;
import java.io.StringReader;
import java.util.Optional;

import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

import de.monticore.commonexpressions._ast.ASTArguments;
import de.monticore.expressions.prettyprint.CommonExpressionsPrettyPrinter;
import de.monticore.expressionsbasis._ast.ASTExpression;
import de.monticore.testcommonexpressions._ast.ASTPrimaryExpression;
import de.monticore.testcommonexpressions._parser.TestCommonExpressionsParser;
import de.monticore.testcommonexpressions._visitor.TestCommonExpressionsVisitor;
import de.se_rwth.commons.logging.Log;
import de.se_rwth.commons.logging.LogStub;

/**
 * @author npichler
 */

public class CommonExpressionsPrettyPrinterTest {
  
  @BeforeClass
  public static void init() {
    LogStub.init();
    Log.enableFailQuick(false);
  }
  
  @Before
  public void setUp() {
    Log.getFindings().clear();
  }
  
  static class PrimaryPrettyPrinter extends CommonExpressionsPrettyPrinter
      implements TestCommonExpressionsVisitor {
    
    private TestCommonExpressionsVisitor realThis;
    
    @Override
    public void visit(ASTPrimaryExpression node) {
      getPrinter().print((node.getName()));
    }
    
    public PrimaryPrettyPrinter(IndentPrinter printer) {
      super(printer);
      realThis = this;
    }
    
    @Override
    public TestCommonExpressionsVisitor getRealThis() {
      return realThis;
    }
  }
  
  @Test
  public void testInfixPrettyPrintAnd() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a && b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintOr() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a || b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintAndOr() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a && b || c"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));  }
  
  @Test
  public void testInfixPrettyPrintMult() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a * b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintDivide() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a / b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintPlus() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a + b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintMinus() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a - b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintLessEquals() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a <= b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintGreaterEquals() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a >= b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintGreaterThan() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a > b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintLessThan() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a < b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintEquals() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a == b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintNotEquals() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a != b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintSimpleAssignment() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a += b"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testInfixPrettyPrintCalculation() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("a + b -c * d"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testBracketExpression() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("(exp)"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testBooleanNotExpression() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("~exp"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testLogicalNotExpression() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("!exp"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testArguments() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTArguments> ast = parser.parseArguments(new StringReader("(a,b,c)"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTArguments assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseArguments(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
  @Test
  public void testCallExpression() throws IOException {
    TestCommonExpressionsParser parser = new TestCommonExpressionsParser();
    Optional<ASTExpression> ast = parser.parseExpression(new StringReader("exp(a,b,c)"));
    assertTrue(ast.isPresent());
    assertFalse(parser.hasErrors());
    ASTExpression assignment = ast.get();
    PrimaryPrettyPrinter printer = new PrimaryPrettyPrinter(new IndentPrinter());
    String output = printer.prettyprint(ast.get());
    ast = parser.parseExpression(new StringReader(output));
    assertFalse(parser.hasErrors());
    assertTrue(ast.isPresent());
    assertTrue(assignment.deepEquals(ast.get()));
  }
  
}
